day 09 functionality

+++ src/main/java/com/xebia/fs101/writerpad/service/ArticleService.java
@@ -3,6 +3,7 @@ package com.xebia.fs101.writerpad.service;
 import com.xebia.fs101.writerpad.entity.Article;
 import com.xebia.fs101.writerpad.entity.User;
 import com.xebia.fs101.writerpad.exception.ArticleNotFoundException;
+import com.xebia.fs101.writerpad.exception.UserNotAuthorizedException;
 import com.xebia.fs101.writerpad.model.ArticleStatus;
 import com.xebia.fs101.writerpad.model.ReadingTime;
 import com.xebia.fs101.writerpad.repository.ArticleRepository;
@@ -58,17 +59,24 @@ public class ArticleService {
                 ArticleNotFoundException::new);
     }
 
-    public void delete(String slugId) {
+    public void delete(String slugId, User user) {
 
         UUID id = StringUtils.extractUuid(slugId);
-        articleRepository.findById(id).orElseThrow(ArticleNotFoundException::new);
+        Article article = articleRepository.findById(id).orElseThrow(
+                ArticleNotFoundException::new);
+        if (!article.getUser().getUsername().equals(user.getUsername()))
+            throw new UserNotAuthorizedException(
+                    "User not athorized to delte the article");
         articleRepository.deleteById(id);
     }
 
-    public Article update(String slugId, Article copyFrom) {
+    public Article update(String slugId, Article copyFrom, User user) {
 
         Article article = articleRepository.findById(StringUtils.extractUuid(slugId))
                 .orElseThrow(ArticleNotFoundException::new);
+        if (!article.getUser().getUsername().equals(user.getUsername()))
+            throw new UserNotAuthorizedException(
+                    "User not authorized to update the article");
         Article articleToBeUpdated = article.update(copyFrom);
         return articleRepository.save(articleToBeUpdated);
     }
